+++ src/lib/components/output.jsx
@@ -5,147 +5,12 @@ import Text from 'react-canvas/lib/Text'
 import FontFace from 'react-canvas/lib/FontFace'
 import Share from './share.jsx!'
 import measureText from 'react-canvas/lib/measureText'
+import Typesetter from '../models/typesetter'
 import './output.scss!post-css'
-import F from 'fkit'
 
 class Line extends React.Component {
   render() {
     return <Text style={this.props.line.style}>{this.props.line.line}</Text>
-  }
-}
-
-let getFontFace = (font) => {
-  let options = {weight: font.weight}
-  if (font.italic) options.style = 'italic'
-  return FontFace(font.name, null, options)
-}
-class LineMetrics {
-  constructor(ctx, width, font, text) {
-    this.text = text
-    this.width = width
-    this.fontFace = getFontFace(font)
-    let fontExpr = size => this.fontFace.attributes.style + ' normal ' + this.fontFace.attributes.weight + ' ' + size + 'pt ' + this.fontFace.family
-    ctx.font = fontExpr(18)
-    let naturalWidth = ctx.measureText(text).width
-    this.fontSize = Math.min(300, 18 * width / naturalWidth)
-    this.height = Math.floor(this.fontSize * 3)
-
-    ctx.font = fontExpr(this.fontSize)
-    ctx.clearRect(0, 0, this.width, this.height)
-    ctx.fillStyle = "black"
-    ctx.textAlign = 'center'
-    ctx.fillText(text, this.width / 2, this.height / 2, this.width)
-    this.calculateDepthMap(ctx.getImageData(0, 0, this.width, this.height))
-  }
-
-  calculateDepthMap(imageData) {
-    let pixels = new Uint32Array(imageData.data.buffer),
-      topBuffer = new ArrayBuffer(this.width * 4),
-      bottomBuffer = new ArrayBuffer(this.width * 4)
-      this.topDepth = new Uint32Array(topBuffer)
-      this.bottomDepth = new Uint32Array(bottomBuffer)
-
-    for (var line = 0; line < this.height; line++) {
-      for (var col = 0; col < this.width; col++) {
-        if (!this.topDepth[col]) {
-          var topCellIdx = line * this.width + col
-          if (pixels[topCellIdx] !== 0) this.topDepth[col] = line
-        }
-        if (!this.bottomDepth[col]) {
-          var bottomCellIdx = (this.height - line - 1) * this.width + col
-          if (pixels[bottomCellIdx] !== 0) this.bottomDepth[col] = line
-        }
-      }
-    }
-  }
-
-  getMinDepth(depth) {
-    let min = this.height / 2
-    for (var i = 0; i < depth.length; i++) {
-      var d = depth[i]
-      if (d > 0 && d < min) min = d
-    }
-    return min
-  }
-  getLeadingFromTop() {
-    return this.getMinDepth(this.topDepth)
-  }
-  getLeadingFromBottom() {
-    return this.getMinDepth(this.bottomDepth)
-  }
-
-  getLeading(prevMetrics) {
-    let min = this.height / 2 + prevMetrics.height / 2
-    for (var i = 0; i < this.topDepth.length; i++) {
-      var d1 = this.topDepth[i], d2 = prevMetrics.bottomDepth[i]
-      if (d1 > 0 && d2 > 0 && (d1 + d2) < min) min = d1 + d2
-    }
-    return min
-  }
-}
-
-class Typesetter {
-  constructor(typePair, width, spacing) {
-    this.typePair = typePair
-    this.width = width
-    this.spacing = spacing
-    this.metricsCache = new Map()
-    this.setupCanvas()
-  }
-
-  setupCanvas() {
-    this.canvas = document.querySelector('body > canvas') || document.querySelector("body").appendChild(document.createElement("canvas"))
-    this.canvas.style.backgroundColor = "palegoldenrod"
-    this.canvas.width = this.width
-    this.canvas.height = 900
-    this.ctx = this.canvas.getContext("2d")
-  }
-
-  getMetrics(line) {
-    if (!this.metricsCache.has(line)) {
-      let font = this.typePair.main, text = line;
-      if (line.match(/^!/)) {
-        font = this.typePair.alt;
-        text = text.replace(/^!/, '');
-      }
-      text = font.caps ? text.toUpperCase() : text
-      this.metricsCache.set(line, new LineMetrics(this.ctx, this.width, font, text))
-    }
-    return this.metricsCache.get(line)
-  }
-
-  setLines(lines, chosenColor) {
-    let linesWithMetrics = lines.map(line => this.getMetrics(line))
-    let totalHeight = this.spacing
-    let sizedLines = []
-    for (var i = 0; i < linesWithMetrics.length; i++) {
-      let line = linesWithMetrics[i], text = line.text
-      if (i == 0) {
-        totalHeight += line.height / 2 - line.getLeadingFromTop()
-      } else {
-        let prev = linesWithMetrics[i-1]
-        //totalHeight += prev.height / 2 - prev.getLeadingFromBottom()
-        //totalHeight += line.height / 2 - line.getLeadingFromTop()
-        totalHeight += this.spacing / 4 + (prev.height + line.height) / 2 - line.getLeading(prev)
-      }
-      let style = {
-          fontSize: line.fontSize,
-          height: line.height,
-          lineHeight: 0,
-          top: totalHeight - line.fontSize,
-          width: this.width + this.spacing * 2,
-          fontFace: line.fontFace,
-          left: 0,
-          textAlign: 'center',
-          color: chosenColor.foreground,
-          zIndex: 2
-        }
-      if (i == linesWithMetrics.length - 1) {
-        totalHeight += line.height / 2 - line.getLeadingFromBottom()
-      }
-      sizedLines.push({line: text, style})
-    }
-    return {totalHeight, sizedLines}
   }
 }
